extend and move MLSController mock
